--- entregas/entrega-001/Azul.java
@@ -1,7 +1,0 @@
-public class Azul extends Queso {
-    public Azul() {
-        super("Queso Azul");
-        this.forma = "-.-.-.-.-.-.";
-    }
-
-}
