--- entregas/entrega-001/Brioche.java
@@ -1,8 +1,0 @@
-public class Brioche extends Pan {
-
-    public Brioche(){
-        super("Brioche");
-        this.forma = "############";
-
-    }
-}
