--- entregas/entrega-001/Cheddar.java
@@ -1,6 +1,0 @@
-public class Cheddar extends Queso {
-    public Cheddar() {
-        super("Queso Cheddar");
-        this.forma = "============";
-    }
-}
