--- entregas/entrega-001/Extras.java
@@ -1,5 +1,0 @@
-public abstract class Extras extends Ingrediente {
-    public Extras(String nombre){
-        this.nombre = nombre;
-    }
-}
