--- entregas/entrega-001/Huevo.java
@@ -1,6 +1,0 @@
-public class Huevo extends Extras{
-    public Huevo(){
-        super("Huevos");
-        this.forma = "~-~-(   )~-~-";
-    }
-}
