--- entregas/entrega-001/Ingrediente.java
@@ -1,12 +1,0 @@
-public abstract class Ingrediente {
-    protected String nombre;
-    protected String forma;
-
-    public String describir(){
-        return nombre;
-    }
-
-    public String mostrar(){
-        return forma;
-    }
-}
