--- entregas/entrega-001/Integral.java
@@ -1,7 +1,0 @@
-public class Integral extends Pan {
-
-    public Integral(){
-        super("Integral");
-        this.forma = "############";
-    }
-}
