--- entregas/entrega-001/Ketchup.java
@@ -1,8 +1,0 @@
-public class Ketchup extends Salsa {
-
-    public Ketchup(String cantidad){
-        super("Salsa Ketchup");
-        this.forma = "ooOooOooOooO";
-        this.cantidad = cantidad;
-    }
-}
