--- entregas/entrega-001/Mayonesa.java
@@ -1,8 +1,0 @@
-public class Mayonesa extends Salsa {
-
-    public Mayonesa(String cantidad){
-        super("Salsa Mayonesa");
-        this.forma = "~=~=~=~=~=~=";
-        this.cantidad = cantidad;
-    }
-}
