--- entregas/entrega-001/Pan.java
@@ -1,5 +1,0 @@
-public abstract class Pan extends Ingrediente {
-    public Pan(String nombre) {
-        this.nombre = nombre;
-    }
-}
