--- entregas/entrega-001/Pollo.java
@@ -1,6 +1,0 @@
-public class Pollo extends Carne {
-    public Pollo(String coccion) {
-        super("Pollo", coccion);
-        this.forma =  ";;;;;;;;;;;;";
-    }
-}
