--- entregas/entrega-001/Queso.java
@@ -1,5 +1,0 @@
-public abstract class Queso extends Extras {
-    public Queso(String nombre){
-        super(nombre);
-    }
-}
