--- entregas/entrega-001/Salsa.java
@@ -1,9 +1,0 @@
-public abstract class Salsa extends Extras{
-    protected String cantidad;
-    public Salsa(String nombre){
-        super(nombre);
-    }
-    public String describir(){
-        return cantidad + " de " + nombre;
-    }
-}
