+++ entregas/tresgalloRuben/entrega-001/Azul.java
@@ -1,0 +1,7 @@
+public class Azul extends Queso {
+    public Azul() {
+        super("Queso Azul");
+        this.forma = "-.-.-.-.-.-.";
+    }
+
+}
