+++ entregas/tresgalloRuben/entrega-001/Brioche.java
@@ -1,0 +1,8 @@
+public class Brioche extends Pan {
+
+    public Brioche(){
+        super("Brioche");
+        this.forma = "############";
+
+    }
+}
