+++ entregas/tresgalloRuben/entrega-001/Cheddar.java
@@ -1,0 +1,6 @@
+public class Cheddar extends Queso {
+    public Cheddar() {
+        super("Queso Cheddar");
+        this.forma = "============";
+    }
+}
