+++ entregas/tresgalloRuben/entrega-001/Cocinero.java
@@ -1,0 +1,19 @@
+class Cocinero extends Empleado {
+
+    public void prepararHamburguesa() {
+        final int NUMERO_EXTRAS = 4;
+        Carne laCarne = new Res("poco hecha");
+        hamburguesa = new Hamburguesa(new Brioche(), new Brioche(), laCarne, NUMERO_EXTRAS);
+
+        hamburguesa.agregarExtra(new Cheddar());
+        hamburguesa.agregarExtra(new Ketchup("poco"));
+        hamburguesa.agregarExtra(new Mayonesa("mucho"));
+        hamburguesa.agregarExtra(new Huevo());
+        this.tieneHamburguesa = true;
+    }
+
+    public Hamburguesa entregar() {
+        this.tieneHamburguesa = false;
+        return hamburguesa;
+    }
+}
