+++ entregas/tresgalloRuben/entrega-001/Empleado.java
@@ -1,0 +1,8 @@
+abstract class Empleado {
+    protected Hamburguesa hamburguesa;
+    protected boolean tieneHamburguesa;
+
+    public Empleado() {
+        this.tieneHamburguesa = false;
+    }
+}
