+++ entregas/tresgalloRuben/entrega-001/Extras.java
@@ -1,0 +1,5 @@
+public abstract class Extras extends Ingrediente {
+    public Extras(String nombre){
+        this.nombre = nombre;
+    }
+}
