+++ entregas/tresgalloRuben/entrega-001/Huevo.java
@@ -1,0 +1,6 @@
+public class Huevo extends Extras{
+    public Huevo(){
+        super("Huevos");
+        this.forma = "~-~-(   )~-~-";
+    }
+}
