+++ entregas/tresgalloRuben/entrega-001/Ingrediente.java
@@ -1,0 +1,12 @@
+public abstract class Ingrediente {
+    protected String nombre;
+    protected String forma;
+
+    public String describir(){
+        return nombre;
+    }
+
+    public String mostrar(){
+        return forma;
+    }
+}
