+++ entregas/tresgalloRuben/entrega-001/Integral.java
@@ -1,0 +1,7 @@
+public class Integral extends Pan {
+
+    public Integral(){
+        super("Integral");
+        this.forma = "############";
+    }
+}
