+++ entregas/tresgalloRuben/entrega-001/Ketchup.java
@@ -1,0 +1,8 @@
+public class Ketchup extends Salsa {
+
+    public Ketchup(String cantidad){
+        super("Salsa Ketchup");
+        this.forma = "ooOooOooOooO";
+        this.cantidad = cantidad;
+    }
+}
