+++ entregas/tresgalloRuben/entrega-001/Mayonesa.java
@@ -1,0 +1,8 @@
+public class Mayonesa extends Salsa {
+
+    public Mayonesa(String cantidad){
+        super("Salsa Mayonesa");
+        this.forma = "~=~=~=~=~=~=";
+        this.cantidad = cantidad;
+    }
+}
