+++ entregas/tresgalloRuben/entrega-001/Pan.java
@@ -1,0 +1,5 @@
+public abstract class Pan extends Ingrediente {
+    public Pan(String nombre) {
+        this.nombre = nombre;
+    }
+}
