+++ entregas/tresgalloRuben/entrega-001/Pollo.java
@@ -1,0 +1,6 @@
+public class Pollo extends Carne {
+    public Pollo(String coccion) {
+        super("Pollo", coccion);
+        this.forma =  ";;;;;;;;;;;;";
+    }
+}
