+++ entregas/tresgalloRuben/entrega-001/Queso.java
@@ -1,0 +1,5 @@
+public abstract class Queso extends Extras {
+    public Queso(String nombre){
+        super(nombre);
+    }
+}
