+++ entregas/tresgalloRuben/entrega-001/Res.java
@@ -1,0 +1,8 @@
+public class Res extends Carne {
+    public Res(String coccion) {
+        super("Res", coccion);
+        this.forma = "::::::::::::";
+    }
+
+
+}
