+++ entregas/tresgalloRuben/reto0710/Node.java
@@ -1,4 +1,4 @@
-package entregas.reto002;
+package entregas.tresgalloRuben.reto0710;
 
 public class Node {
